examples/python/requirements.txt
httpx
web3
python-dotenv
